+++ lib/src/DataModel.dart
@@ -1,5 +1,6 @@
+export 'APIException.dart' show APIException;
 export 'Book.dart' show Book, BookTitle;
-export 'Exceptions.dart' show APIException;
 export 'Image.dart' show Image;
 export 'Search.dart' show Search, SearchSort;
 export 'Tag.dart' show Tag, TagType;
+export 'TagsList.dart' show TagsList;